+++ internal/models/order.go
@@ -1,6 +1,6 @@
 package models
 
 type Order struct {
-	OrderID int `db:"order_id" json:"order_id"`
-	UserID  int `db:"user_id" json:"user_id"`
+	OrderID int   `db:"order_id" json:"order_id"`
+	UserID  int64 `db:"user_id" json:"user_id"`
 }
